+++ Classes/Private/Manager/OCTManager.m
@@ -26,7 +26,7 @@
 @interface OCTManager () <OCTToxDelegate, OCTSubmanagerDataSource>
 
 @property (strong, nonatomic, readonly) OCTTox *tox;
-@property (copy, nonatomic, readonly) OCTManagerConfiguration *configuration;
+@property (copy, nonatomic, readonly) OCTManagerConfiguration *currentConfiguration;
 
 @property (strong, nonatomic, readwrite) OCTSubmanagerAvatars *avatars;
 @property (strong, nonatomic, readwrite) OCTSubmanagerBootstrap *bootstrap;
@@ -60,7 +60,7 @@
     }
 
     [self validateConfiguration:configuration];
-    _configuration = [configuration copy];
+    _currentConfiguration = [configuration copy];
 
     if (! [self importToxSaveIfNeeded]) {
         [self fillError:error withInitErrorCode:OCTManagerInitErrorCannotImportToxSave];
@@ -69,12 +69,10 @@
 
     NSData *savedData = [self getSavedData];
 
-    if (! [self createEncryptSaveWithPassphrase:_configuration.passphrase toxData:savedData]) {
+    if (! [self createEncryptSaveWithPassphrase:_currentConfiguration.passphrase toxData:savedData]) {
         [self fillError:error withInitErrorCode:OCTManagerInitErrorPassphraseFailed];
         return nil;
     }
-    // no need to keep passphrase in memory.
-    _configuration.passphrase = nil;
 
     BOOL wasDecryptError = NO;
     savedData = [self decryptSavedDataIfNeeded:savedData error:error wasDecryptError:&wasDecryptError];
@@ -101,11 +99,16 @@
 
 #pragma mark -  Public
 
+- (OCTManagerConfiguration *)configuration
+{
+    return [self.currentConfiguration copy];
+}
+
 - (NSString *)exportToxSaveFile:(NSError **)error
 {
     @synchronized(self.toxSaveFileLock) {
-        NSString *savedDataPath = self.configuration.fileStorage.pathForToxSaveFile;
-        NSString *tempPath = self.configuration.fileStorage.pathForTemporaryFilesDirectory;
+        NSString *savedDataPath = self.currentConfiguration.fileStorage.pathForToxSaveFile;
+        NSString *tempPath = self.currentConfiguration.fileStorage.pathForTemporaryFilesDirectory;
         tempPath = [tempPath stringByAppendingPathComponent:[savedDataPath lastPathComponent]];
 
         NSFileManager *fileManager = [NSFileManager defaultManager];
@@ -124,6 +127,8 @@
 
 - (BOOL)changePassphrase:(NSString *)passphrase
 {
+    self.currentConfiguration.passphrase = passphrase;
+
     // Passing nil as tox data as we are setting new passphrase.
     BOOL result = [self createEncryptSaveWithPassphrase:passphrase toxData:nil];
 
@@ -156,7 +161,7 @@
 
 - (id<OCTFileStorageProtocol>)managerGetFileStorage
 {
-    return self.configuration.fileStorage;
+    return self.currentConfiguration.fileStorage;
 }
 
 - (NSNotificationCenter *)managerGetNotificationCenter
@@ -200,9 +205,9 @@
 {
     NSFileManager *fileManager = [NSFileManager defaultManager];
 
-    if (_configuration.importToxSaveFromPath && [fileManager fileExistsAtPath:_configuration.importToxSaveFromPath]) {
-        return [fileManager copyItemAtPath:_configuration.importToxSaveFromPath
-                                    toPath:_configuration.fileStorage.pathForToxSaveFile
+    if (_currentConfiguration.importToxSaveFromPath && [fileManager fileExistsAtPath:_currentConfiguration.importToxSaveFromPath]) {
+        return [fileManager copyItemAtPath:_currentConfiguration.importToxSaveFromPath
+                                    toPath:_currentConfiguration.fileStorage.pathForToxSaveFile
                                      error:nil];
     }
 
@@ -211,7 +216,7 @@
 
 - (NSData *)getSavedData
 {
-    NSString *savedDataPath = _configuration.fileStorage.pathForToxSaveFile;
+    NSString *savedDataPath = _currentConfiguration.fileStorage.pathForToxSaveFile;
 
     return [[NSFileManager defaultManager] fileExistsAtPath:savedDataPath] ?
            [NSData dataWithContentsOfFile : savedDataPath] :
@@ -256,7 +261,7 @@
 {
     NSError *toxError = nil;
 
-    _tox = [[OCTTox alloc] initWithOptions:_configuration.options savedData:savedData error:&toxError];
+    _tox = [[OCTTox alloc] initWithOptions:_currentConfiguration.options savedData:savedData error:&toxError];
 
     if (_tox) {
         _toxSaveFileLock = [NSObject new];
@@ -308,7 +313,7 @@
 
 - (void)createRealmManager
 {
-    _realmManager = [[OCTRealmManager alloc] initWithDatabasePath:_configuration.fileStorage.pathForDatabase];
+    _realmManager = [[OCTRealmManager alloc] initWithDatabasePath:_currentConfiguration.fileStorage.pathForDatabase];
 }
 
 - (void)createSubmanagers
@@ -461,7 +466,7 @@
             }
         }
 
-        if (! [data writeToFile:self.configuration.fileStorage.pathForToxSaveFile options:NSDataWritingAtomic error:&error]) {
+        if (! [data writeToFile:self.currentConfiguration.fileStorage.pathForToxSaveFile options:NSDataWritingAtomic error:&error]) {
             throwException(error);
         }
     }